requirements.txt
streamlit
transformers
torch
pillow
requests